refactor: tests controllers refactored
