+++ backend/src/controllers/tests/getTestResult/utils/mappers.ts
@@ -2,7 +2,7 @@ import { PERCENTS, TWO_DIGITS } from 'config/constants';
 import { IClientCourse } from 'interfaces/Ientities/IclientCourses';
 import { ITestResultResponse } from 'interfaces/IResponse/IResponse';
 
-const normaliseTestResult = (testResult: IClientCourse['testResult']): ITestResultResponse => {
+const mapTestResult = (testResult: IClientCourse['testResult']): ITestResultResponse => {
   const answers = testResult.map((question) => ({ qN: question.qN, isCorrect: question.aN }));
   const percentage =
     parseFloat(
@@ -14,4 +14,4 @@ const normaliseTestResult = (testResult: IClientCourse['testResult']): ITestResu
   return response;
 };
 
-export { normaliseTestResult };
+export { mapTestResult };
