+++ backend/src/controllers/tests/index.ts
@@ -1,0 +1,7 @@
+export { default as getAchievements } from './getAchievements/getAchievements';
+export { default as getTest } from './getTest/getTest';
+export { default as getTestResult } from './getTestResult/getTestResult';
+export { default as getTestTime } from './getTestTime/getTestTime';
+export { default as passTest } from './passTest/passTest';
+export { default as unitTestResults } from './sendTestResults/sendTestResults';
+export { default as startTest } from './startTest/startTest';
