+++ backend/src/controllers/tests/startTest/utils/helpers.ts
@@ -1,0 +1,10 @@
+import moment from 'moment';
+
+const generateStartAndFinishTestDates = (testTimeSeconds: number): [Date, Date] => {
+  const startTestDate = new Date();
+  const finishTestDate = moment(startTestDate).add(testTimeSeconds, 's').toDate();
+
+  return [startTestDate, finishTestDate];
+};
+
+export { generateStartAndFinishTestDates };
