+++ backend/src/utils/date/testDate.ts
@@ -3,13 +3,6 @@ import moment from 'moment';
 import { DISABLE_TEST_DAYS } from 'config/constants';
 import { UPDATE_TEST_STATUSES_JOB_START_HOURS } from 'utils/schedule/constants';
 
-const generateStartAndFinishTestDates = (testTimeSeconds: number): [Date, Date] => {
-  const startTestDate = new Date();
-  const finishTestDate = moment(startTestDate).add(testTimeSeconds, 's').toDate();
-
-  return [startTestDate, finishTestDate];
-};
-
 const generateTestStatusToUpdateDates = (checkRangeHours: number): [Date, Date] => {
   const currentDate = new Date();
 
@@ -31,4 +24,4 @@ const generateTestStatusToUpdateDates = (checkRangeHours: number): [Date, Date] 
   return [from, to];
 };
 
-export { generateStartAndFinishTestDates, generateTestStatusToUpdateDates };
+export { generateTestStatusToUpdateDates };
